Conflicts:
	frappe/commands.py
	frappe/core/doctype/docfield/docfield.json
	frappe/core/doctype/docperm/docperm.json
	frappe/core/doctype/patch_log/patch_log.json
	frappe/core/doctype/scheduler_log/scheduler_log.json
	frappe/core/doctype/version/version.json
	frappe/custom/doctype/custom_field/custom_field.json
	frappe/custom/doctype/customize_form_field/customize_form_field.json
	frappe/desk/page/setup_wizard/setup_wizard.js
	frappe/desk/page/setup_wizard/setup_wizard.py
	frappe/docs/assets/img/generators.png
	frappe/docs/user/en/guides/portal-development/generators.md
	frappe/docs/user/fr/guides/app-development/using-html-templates-in-javascript.md
	frappe/docs/user/fr/guides/basics/site_config.md
	frappe/email/doctype/email_account/email_account.py
	frappe/email/doctype/email_account/test_email_account.py
	frappe/email/doctype/email_alert_recipient/email_alert_recipient.json
	frappe/email/queue.py
	frappe/hooks.py
	frappe/integrations/doctype/dropbox_backup/dropbox_backup.py
	frappe/model/base_document.py
	frappe/model/document.py
	frappe/patches.txt
	frappe/public/js/frappe/form/grid.js
	frappe/public/js/frappe/ui/page.js
	frappe/public/js/legacy/clientscriptAPI.js
	frappe/templates/pages/update-password.html
	frappe/translations/ar.csv
	frappe/translations/bg.csv
	frappe/translations/bn.csv
	frappe/translations/bs.csv
	frappe/translations/ca.csv
	frappe/translations/cs.csv
	frappe/translations/da-DK.csv
	frappe/translations/da.csv
	frappe/translations/de.csv
	frappe/translations/el.csv
	frappe/translations/es-PE.csv
	frappe/translations/es.csv
	frappe/translations/et.csv
	frappe/translations/fa.csv
	frappe/translations/fi.csv
	frappe/translations/fr.csv
	frappe/translations/gu.csv
	frappe/translations/he.csv
	frappe/translations/hi.csv
	frappe/translations/hr.csv
	frappe/translations/hu.csv
	frappe/translations/id.csv
	frappe/translations/it.csv
	frappe/translations/ja.csv
	frappe/translations/km.csv
	frappe/translations/kn.csv
	frappe/translations/ko.csv
	frappe/translations/lv.csv
	frappe/translations/mk.csv
	frappe/translations/ml.csv
	frappe/translations/mr.csv
	frappe/translations/ms.csv
	frappe/translations/my.csv
	frappe/translations/nl.csv
	frappe/translations/no.csv
	frappe/translations/pl.csv
	frappe/translations/pt-BR.csv
	frappe/translations/pt.csv
	frappe/translations/ro.csv
	frappe/translations/ru.csv
	frappe/translations/sk.csv
	frappe/translations/sl.csv
	frappe/translations/sq.csv
	frappe/translations/sr.csv
	frappe/translations/sv.csv
	frappe/translations/ta.csv
	frappe/translations/te.csv
	frappe/translations/th.csv
	frappe/translations/tr.csv
	frappe/translations/uk.csv
	frappe/translations/ur.csv
	frappe/translations/vi.csv
	frappe/translations/zh-cn.csv
	frappe/translations/zh-tw.csv
	frappe/utils/__init__.py
	frappe/utils/setup_docs.py
	frappe/website/doctype/blog_category/blog_category.json
	frappe/website/doctype/blog_post/blog_post.json
	frappe/website/doctype/blog_post/blog_post.py
	frappe/website/doctype/web_form/web_form.json
	frappe/website/doctype/web_form/web_form.py
	frappe/website/doctype/web_form_field/web_form_field.json
	frappe/website/doctype/web_page/web_page.json
	frappe/website/doctype/web_page/web_page.py
	frappe/website/doctype/website_theme/website_theme.json
	frappe/workflow/doctype/workflow_transition/workflow_transition.json
	frappe/www/__init__.py
	frappe/www/profile.html
	setup.py

+++ frappe/model/delete_doc.py
@@ -8,6 +8,7 @@ import frappe.model.meta
 from frappe.model.dynamic_links import get_dynamic_link_map
 import frappe.defaults
 from frappe.utils.file_manager import remove_all
+from frappe.utils.password import delete_all_passwords_for
 from frappe import _
 from frappe.model.naming import revert_series_if_last
 
@@ -36,6 +37,9 @@ def delete_doc(doctype=None, name=None, force=0, ignore_doctypes=None, for_reloa
 		# delete attachments
 		remove_all(doctype, name)
 
+		# delete passwords
+		delete_all_passwords_for(doctype, name)
+
 		doc = None
 		if doctype=="DocType":
 			if for_reload:
@@ -48,6 +52,11 @@ def delete_doc(doctype=None, name=None, force=0, ignore_doctypes=None, for_reloa
 					doc.run_method("before_reload")
 
 			else:
+				doc = frappe.get_doc(doctype, name)
+				
+				update_flags(doc, flags, ignore_permissions)
+				check_permission_and_not_submitted(doc)
+					
 				frappe.db.sql("delete from `tabCustom Field` where dt = %s", name)
 				frappe.db.sql("delete from `tabCustom Script` where dt = %s", name)
 				frappe.db.sql("delete from `tabProperty Setter` where doc_type = %s", name)
@@ -59,13 +68,7 @@ def delete_doc(doctype=None, name=None, force=0, ignore_doctypes=None, for_reloa
 			doc = frappe.get_doc(doctype, name)
 
 			if not for_reload:
-				if ignore_permissions:
-					if not flags: flags = {}
-					flags["ignore_permissions"] = ignore_permissions
-
-				if flags:
-					doc.flags.update(flags)
-
+				update_flags(doc, flags, ignore_permissions)
 				check_permission_and_not_submitted(doc)
 
 				if not ignore_on_trash:
@@ -136,10 +139,18 @@ def delete_from_table(doctype, name, ignore_doctypes, doc):
 	for t in list(set(tables)):
 		if t not in ignore_doctypes:
 			frappe.db.sql("delete from `tab%s` where parenttype=%s and parent = %s" % (t, '%s', '%s'), (doctype, name))
+			
+def update_flags(doc, flags=None, ignore_permissions=False):
+	if ignore_permissions:
+		if not flags: flags = {}
+		flags["ignore_permissions"] = ignore_permissions
+
+	if flags:
+		doc.flags.update(flags)	
 
 def check_permission_and_not_submitted(doc):
 	# permission
-	if not doc.flags.ignore_permissions and frappe.session.user!="Administrator" and not doc.has_permission("delete"):
+	if not doc.flags.ignore_permissions and frappe.session.user!="Administrator" and (not doc.has_permission("delete") or (doc.doctype=="DocType" and not doc.custom)):
 		frappe.msgprint(_("User not allowed to delete {0}: {1}").format(doc.doctype, doc.name), raise_exception=True)
 
 	# check if submitted
